code correction to free rbuffer in case the open, read and write operation fails

+++ 0x15-file_io/0-read_textfile.c
@@ -15,7 +15,7 @@
 
 ssize_t read_textfile(const char *filename, size_t letters)
 {
-	size_t fd, lr, lw;
+	int fd, lr, lw;
 	char *rbuffer;
 
 	if (filename == NULL)
